requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
